refactor: refactored views of installer
