+++ public/installer/views/layout-bottom.php
@@ -1,0 +1,7 @@
+
+<!-- top layout here -->
+
+<!-- any middle view here -->
+
+    </body>
+</html>
